requirements.txt
tensorflow=2.10.0
numpy=1.26.4
Pillow=11.2.1
matplotlib=3.10.3
opencv-python=4.11.0.86